requirements.txt
coverage==4.4.1
Django==1.11.4
django-bootstrap-toolkit==2.15.0
django-registration-redux==1.7
Jinja2==2.9.6
MarkupSafe==1.0
olefile==0.44
Pillow==4.2.1
psycopg2==2.7.3
pytz==2017.2
